Crate air_interpreter_data

Modules

verification

Structs

ApResult

Describes result of applying functor apply to streams.

ApResultResolver

The resolver for an archived ApResult

ArchivedApResult

An archived ApResult

ArchivedCallServiceFailed

An archived CallServiceFailed

ArchivedCanonCidAggregate

An archived CanonCidAggregate

ArchivedCanonResultCidAggregate

An archived CanonResultCidAggregate

ArchivedCidInfo

An archived CidInfo

ArchivedCidStore

An archived CidStore

ArchivedExecutionTrace

An archived ExecutionTrace

ArchivedFoldResult

An archived FoldResult

ArchivedFoldSubTraceLore

An archived FoldSubTraceLore

ArchivedGenerationIdx

An archived GenerationIdx

ArchivedInterpreterData

An archived InterpreterData

ArchivedParResult

An archived ParResult

ArchivedRawValue

An archived RawValue

ArchivedServiceResultCidAggregate

An archived ServiceResultCidAggregate

ArchivedSubTraceDesc

An archived SubTraceDesc

ArchivedTracePos

An archived TracePos

CallServiceFailed

CallServiceFailedResolver

The resolver for an archived CallServiceFailed

CanonCidAggregate

The type Canon trace CID refers to.

CanonCidAggregateResolver

The resolver for an archived CanonCidAggregate

CanonResultCidAggregate

CanonResultCidAggregateResolver

The resolver for an archived CanonResultCidAggregate

CidInfo

CidInfoResolver

The resolver for an archived CidInfo

CidStore

Stores CID to Value corresponance.

CidStoreResolver

The resolver for an archived CidStore

CidTracker

ExecutionTrace

ExecutionTraceResolver

The resolver for an archived ExecutionTrace

FoldResult

FoldResultResolver

The resolver for an archived FoldResult

FoldSubTraceLore

Let’s consider an example of trace that could be produces by the following fold: (fold $stream v (call 1) (call 2) (next v) (call 3) (call 4) ) x Having started with stream with two elements {v1, v2} the resulted trace would looks like [(1) (2)] [(1) (2)] [(3) (4)] [(3) (4)] <— the sequence of call states v1 v2 v2 v1 <–– corresponding values from $stream that the iterable v had at the moment of call

FoldSubTraceLoreResolver

The resolver for an archived FoldSubTraceLore

GenerationIdx

GenerationIdxResolver

The resolver for an archived GenerationIdx

InterpreterData

The AIR interpreter could be considered as a function f(prev_data: InterpreterData, current_data: InterpreterData, … ) -> (result_data: InterpreterData, …). This function receives prev and current data and produces a result data. All these data have the following format.

InterpreterDataEnvelope

An envelope for the AIR interpreter data that makes AIR data version info accessible in a stable way.

InterpreterDataEnvelopeRepr

InterpreterDataResolver

The resolver for an archived InterpreterData

ParResult

ParResultResolver

The resolver for an archived ParResult

RawValue

RawValueResolver

The resolver for an archived RawValue

ServiceResultCidAggregate

A proof of service result execution result.

ServiceResultCidAggregateResolver

The resolver for an archived ServiceResultCidAggregate

SubTraceDesc

Descriptor of a subtrace inside execution trace.

SubTraceDescResolver

The resolver for an archived SubTraceDesc

TracePos

TracePosResolver

The resolver for an archived TracePos

Versions

Enums

ArchivedCallResult

An archived CallResult

ArchivedCanonResult

An archived CanonResult

ArchivedExecutedState

An archived ExecutedState

ArchivedProvenance

An archived Provenance

ArchivedSender

An archived Sender

ArchivedValueRef

An archived ValueRef

CallResult

CallResultResolver

The resolver for an archived CallResult

CanonResult

Contains ids of element that were on a stream at the moment of an appropriate canon call.

CanonResultResolver

The resolver for an archived CanonResult

CidStoreVerificationError

DataDeserializationError

ExecutedState

ExecutedStateResolver

The resolver for an archived ExecutedState

Provenance

ProvenanceResolver

The resolver for an archived Provenance

Sender

SenderResolver

The resolver for an archived Sender

ValueRef

ValueRefResolver

The resolver for an archived ValueRef

Functions

data_version

Type Aliases

FoldLore

This type represents all information in an execution trace about states executed during a fold execution.

InterpreterDataEnvelopeFormat

PosType

TraceLen