Crate air_interpreter_data

Structs

ApResult

Describes result of applying functor apply to streams.

CallServiceFailed

CanonCidAggregate

The type Canon trace CID refers to.

CanonResultCidAggregate

CidInfo

CidStore

Stores CID to Value corresponance.

CidTracker

ExecutionTrace

FoldResult

FoldSubTraceLore

Let’s consider an example of trace that could be produces by the following fold: (fold $stream v (call 1) (call 2) (next v) (call 3) (call 4) ) x Having started with stream with two elements {v1, v2} the resulted trace would looks like [(1) (2)] [(1) (2)] [(3) (4)] [(3) (4)] <— the sequence of call states v1 v2 v2 v1 <–– corresponding values from $stream that the iterable v had at the moment of call

GenerationIdx

InterpreterData

The AIR interpreter could be considered as a function f(prev_data: InterpreterData, current_data: InterpreterData, … ) -> (result_data: InterpreterData, …). This function receives prev and current data and produces a result data. All these data have the following format.

ParResult

ServiceResultCidAggregate

A proof of service result execution result.

SubTraceDesc

Descriptor of a subtrace inside execution trace.

TracePos

Versions

Enums

CallResult

CanonResult

Contains ids of element that were on a stream at the moment of an appropriate canon call.

ExecutedState

Provenance

Sender

ValueRef

Functions

data_version

Type Aliases

FoldLore

This type represents all information in an execution trace about states executed during a fold execution.

PosType

TraceLen